i18n: complete the English translation and let an admin offer Italian and Norwegian
Redone against origin/next. English covered 777 of 1352 keys and now
covers all of them. 565 of the 572 missing strings came across
unchanged from the same work on another base, since the French source
was identical; the other seven are the new RGESN section and a
datasets label.

Placeholder bugs, where a translation had lost an argument the French
passes, so the value never reached the page:

- general.a11y.desc, improveDesc and stateDesc hardcoded a domain
  instead of taking {domain}, in English and in four other locales.
  That parameter is what lets an instance name itself.
- reveal.impacts.summary.consumption dropped {usage} in English, so
  the sentence said "in this conversation" whichever scenario was
  chosen.
- ranking.energy.desc and ranking.energy.views.methodo.3.descs.1 had
  lost their <a> wrappers, leaving Ecologits and GenAI Impact as
  plain text.
- ranking.methodo.desc.2 asked for {lnotebookLinkProps}, which no
  caller passes.

Drop reveal.equivalent: 18 keys, 111 entries across the locale files.
It belongs to the block the LED bulb, phone battery and video
streaming equivalents replaced, and nothing in src reads it.

SEOHead ignored platform_name, so a renamed instance still carried
compar:IA in every tab title, search result and link preview while
its header showed its own name. Every locale held the brand as a
plain token in seo.title and seo.desc, so both take {platformName}
now. The auth context moves to its own module: SEOHead sits in
PageLayout and should not have to pull in the API client, or make
every test that renders a page build a context.

Italian and Norwegian become selectable. Lithuanian and Swedish
already were, just off by default; these two were not compiled into
the bundle at all, which is what the back-office can offer from. All
four stay out of DEFAULT_ENABLED_LOCALES, so nothing changes for an
existing instance until an admin ticks the box. nb_NO.json becomes
nb-NO.json, since the code reaches the page as <html lang="..."> and
nb_NO is not a valid BCP 47 tag. Both locales carried
ranking.ranking.desc translated from a French text since rewritten,
with a link the current message has no argument for; retranslated.

Terminology follows the same rules as before: model vendors are "the
model lab", descriptive prose says "the arena". Legal statements,
form names and the maintenance page keep the brand.
