+++ frontend/src/lib/consumptionSummary.spec.ts
@@ -156,7 +156,7 @@ describe('buildConsumptionSummary', () => {
     [
       'en',
       'Gemma 4 31B is a medium-sized model',
-      'generating 100 tokens used approximately 100 mWh'
+      'generating 100 tokens used about 100 mWh'
     ],
     ['da', 'Gemma 4 31B er en mellemstor model', 'genereringen af 100 tokens cirka 100 mWh'],
     ['sv', 'Gemma 4 31B är en mellanstor modell', 'genereringen av 100 token cirka 100 mWh'],
